Add provisioning profiles
This allows us to return a provision profile which checks to see if the
node type needs some special type of config. If it does it merges this with
the default configuration to avoid lots of cruft in the yaml config although
possibly this will get confusing later tracing what is all going to be used
for a nodeclass. I used deep_merge because I was too lazy to write my own hash
merging function and it doesn't have any dependencies.Y

+++ config.ru
@@ -15,8 +15,19 @@ if not File.exist?(ENV['EVIE_CONFIG'])
   raise ArgumentError, "#{ENV['EVIE_CONFIG']} doesn't exist or couldn't be opened."
 end
 
+
+if ENV['PROFILE_CONFIG'].nil?
+  raise ArgumentError, 'PROFILE_CONFIG env is not defined.'
+end
+
+if not File.exist?(ENV['PROFILE_CONFIG'])
+  raise ArgumentError, "#{ENV['PROFILE_CONFIG']} doesn't exist or couldn't be opened."
+end
+
+
 use Rack::Config do |env|
   env['evie'] = YAML::load(File.open(ENV['EVIE_CONFIG']))
+  env['profile'] = YAML::load(File.open(ENV['PROFILE_CONFIG']))
 end
 
 # Application Routes
